adds get by id endpoint (spring mvc mode)

+++ src/main/java/com/pigmalion/springwebfluxannotationdemo/repositories/IUserRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.pigmalion.springwebfluxannotationdemo.model.User;
 
 
-public interface IUserRepository extends JpaRepository<User, Integer> {
+public interface IUserRepository extends JpaRepository<User, Long> {
 }
